requirements.txt
absl-py==0.9.0
astor==0.8.0
astunparse==1.6.3
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
blinker==1.4
brotlipy==0.7.0
cachetools==4.1.0
certifi==2020.4.5.2
cffi==1.14.0
chainer==6.3.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.20
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
future==0.18.2
gast==0.2.2
get==2019.4.13
gevent==20.9.0
gevent-websocket==0.10.1
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.8.4
greenlet==0.4.17
grpcio==1.27.2
h5py==2.10.0
idna==2.9
imageio==2.8.0
importlib-metadata==1.6.1
ipykernel==5.3.0
ipython @ file:///tmp/build/80754af9/ipython_1592348411427/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
json5==0.9.5
jsonpatch==1.25
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.2.9
jupyterlab-server==1.2.0
Keras==2.4.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
lightgbm==2.3.1
Mako==1.1.3
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.2.2
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
mock==4.0.2
mxnet==1.6.0
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
notebook==5.6.0
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.1.0
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
post==2019.4.13
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.3
ptyprocess==0.6.0
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
pydot==1.4.1
Pygments==2.6.1
pygpu==0.7.6
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==18.1.1
qtconsole==4.7.4
QtPy==1.9.0
query-string==2019.4.13
request==2019.4.13
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
tensorboard==1.15.0
tensorboard-plugin-wit==1.6.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tflearn==0.3.2
Theano==1.0.4
threadpoolctl==2.1.0
tifffile==2020.6.3
torch==1.7.0
torchaudio==0.7.0
torchfile==0.1.0
torchnet @ file:///root/tnt
torchvision==0.6.1
tornado==5.1.1
tqdm==4.46.1
traitlets==4.3.3
typing==3.7.4.1
typing-extensions==3.7.4.1
urllib3==1.25.9
visdom==0.1.8.9
wcwidth==0.2.4
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.16.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xgboost==1.1.1
zipp==3.1.0
zope.event==4.5.0
zope.interface==5.2.0
